restyle homepage carousel
